requirements.txt
streamlit
pandas
chromadb
sentence-transformers
langchain
langchain-groq
python-dotenv
